requirements.txt
accelerate==0.27.2
annotated-types==0.6.0
anyio==4.3.0
charset-normalizer==2.0.4
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
einops==0.7.0
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.3
mkl-fft==1.2.4
mkl-random==1.2.4
mkl-service==2.4.0
networkx==3.2.1
numpy==1.26.3
openai==1.12.0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psutil==5.9.8
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
transformers==4.37.2
transformers-stream-generator==0.0.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3 ==2.1.0
